go.wde/cocoa: moved + updated cocoa framework code for go-gettable build
Moved the cocoa framework code to the cocoa package so that once go is
updated with CL 60590044, the package is go-gettable and all the source
is tracked by the go tool.

As a stepping stone, a Makefile to build the sources into GOARCH
dependent .syso files also is present
so the updated code just needs to run make to build syso’s that the go
tool in 1.2 can link in. The syso’s
may be checked in as part of the final merge, if go 1.3 doesn’t accept
the CL, or is still a ways off.

A porting README will come soon describing the process of updating the
cocoa package and making it
work in a go-gettable way. Also the silly nib hack will be shown as
well.